Speedup string slice decoding.
