+++ types.go
@@ -35,68 +35,6 @@ func Register(typ reflect.Type, enc encoderFunc, dec decoderFunc) {
 
 var structs = newStructCache()
 
-var (
-	valueEncoders []encoderFunc
-	valueDecoders []decoderFunc
-)
-
-func init() {
-	valueEncoders = []encoderFunc{
-		reflect.Bool:          encodeBoolValue,
-		reflect.Int:           encodeInt64Value,
-		reflect.Int8:          encodeInt64Value,
-		reflect.Int16:         encodeInt64Value,
-		reflect.Int32:         encodeInt64Value,
-		reflect.Int64:         encodeInt64Value,
-		reflect.Uint:          encodeUint64Value,
-		reflect.Uint8:         encodeUint64Value,
-		reflect.Uint16:        encodeUint64Value,
-		reflect.Uint32:        encodeUint64Value,
-		reflect.Uint64:        encodeUint64Value,
-		reflect.Float32:       encodeFloat32Value,
-		reflect.Float64:       encodeFloat64Value,
-		reflect.Complex64:     encodeUnsupportedValue,
-		reflect.Complex128:    encodeUnsupportedValue,
-		reflect.Array:         encodeArrayValue,
-		reflect.Chan:          encodeUnsupportedValue,
-		reflect.Func:          encodeUnsupportedValue,
-		reflect.Interface:     encodeInterfaceValue,
-		reflect.Map:           encodeMapValue,
-		reflect.Ptr:           encodeUnsupportedValue,
-		reflect.Slice:         encodeSliceValue,
-		reflect.String:        encodeStringValue,
-		reflect.Struct:        encodeStructValue,
-		reflect.UnsafePointer: encodeUnsupportedValue,
-	}
-	valueDecoders = []decoderFunc{
-		reflect.Bool:          decodeBoolValue,
-		reflect.Int:           decodeInt64Value,
-		reflect.Int8:          decodeInt64Value,
-		reflect.Int16:         decodeInt64Value,
-		reflect.Int32:         decodeInt64Value,
-		reflect.Int64:         decodeInt64Value,
-		reflect.Uint:          decodeUint64Value,
-		reflect.Uint8:         decodeUint64Value,
-		reflect.Uint16:        decodeUint64Value,
-		reflect.Uint32:        decodeUint64Value,
-		reflect.Uint64:        decodeUint64Value,
-		reflect.Float32:       decodeFloat64Value,
-		reflect.Float64:       decodeFloat64Value,
-		reflect.Complex64:     decodeUnsupportedValue,
-		reflect.Complex128:    decodeUnsupportedValue,
-		reflect.Array:         decodeArrayValue,
-		reflect.Chan:          decodeUnsupportedValue,
-		reflect.Func:          decodeUnsupportedValue,
-		reflect.Interface:     decodeInterfaceValue,
-		reflect.Map:           decodeMapValue,
-		reflect.Ptr:           decodeUnsupportedValue,
-		reflect.Slice:         decodeSliceValue,
-		reflect.String:        decodeStringValue,
-		reflect.Struct:        decodeStructValue,
-		reflect.UnsafePointer: decodeUnsupportedValue,
-	}
-}
-
 //------------------------------------------------------------------------------
 
 func encodeUnsupportedValue(e *Encoder, v reflect.Value) error {
@@ -160,14 +98,6 @@ func encodeByteArrayValue(e *Encoder, v reflect.Value) error {
 	b := make([]byte, v.Len())
 	reflect.Copy(reflect.ValueOf(b), v)
 	return e.write(b)
-}
-
-func decodeByteSliceValue(d *Decoder, v reflect.Value) error {
-	return d.byteSliceValue(v)
-}
-
-func decodeByteArrayValue(d *Decoder, v reflect.Value) error {
-	return d.byteArrayValue(v)
 }
 
 //------------------------------------------------------------------------------
@@ -360,96 +290,4 @@ func (m *structCache) Fields(typ reflect.Type) *fields {
 	}
 
 	return fs
-}
-
-func getEncoder(typ reflect.Type) encoderFunc {
-	enc := _getEncoder(typ)
-	if id := extTypeId(typ); id != -1 {
-		return makeExtEncoder(id, enc)
-	}
-	return enc
-}
-
-func _getEncoder(typ reflect.Type) encoderFunc {
-	kind := typ.Kind()
-
-	if typ.Implements(encoderType) {
-		return encodeCustomValue
-	}
-
-	// Addressable struct field value.
-	if reflect.PtrTo(typ).Implements(encoderType) {
-		return encodeCustomValuePtr
-	}
-
-	if typ.Implements(marshalerType) {
-		return marshalValue
-	}
-	if encoder, ok := typEncMap[typ]; ok {
-		return encoder
-	}
-
-	switch kind {
-	case reflect.Ptr:
-		return ptrEncoderFunc(typ)
-	case reflect.Slice:
-		if typ.Elem().Kind() == reflect.Uint8 {
-			return encodeByteSliceValue
-		}
-	case reflect.Array:
-		if typ.Elem().Kind() == reflect.Uint8 {
-			return encodeByteArrayValue
-		}
-	case reflect.Map:
-		if typ.Key().Kind() == reflect.String {
-			switch typ.Elem().Kind() {
-			case reflect.String:
-				return encodeMapStringStringValue
-			case reflect.Interface:
-				return encodeMapStringInterfaceValue
-			}
-		}
-	}
-	return valueEncoders[kind]
-}
-
-func getDecoder(typ reflect.Type) decoderFunc {
-	kind := typ.Kind()
-
-	// Addressable struct field value.
-	if kind != reflect.Ptr && reflect.PtrTo(typ).Implements(decoderType) {
-		return decodeCustomValuePtr
-	}
-
-	if typ.Implements(decoderType) {
-		return decodeCustomValue
-	}
-
-	if typ.Implements(unmarshalerType) {
-		return unmarshalValue
-	}
-
-	if decoder, ok := typDecMap[typ]; ok {
-		return decoder
-	}
-
-	switch kind {
-	case reflect.Ptr:
-		return ptrDecoderFunc(typ)
-	case reflect.Slice:
-		if typ.Elem().Kind() == reflect.Uint8 {
-			return decodeByteSliceValue
-		}
-	case reflect.Array:
-		if typ.Elem().Kind() == reflect.Uint8 {
-			return decodeByteArrayValue
-		}
-	case reflect.Map:
-		if typ.Key().Kind() == reflect.String {
-			if typ.Elem().Kind() == reflect.String {
-				return decodeMapStringStringValue
-			}
-		}
-	}
-	return valueDecoders[kind]
 }
